refactor: address code review feedback for prefab system
1. Extract FindGameObjectByName to shared GameObjectFinder utility class
   - New file: Editor/Scripts/Utilities/GameObjectFinder.cs
   - Removes code duplication between CreatePrefabHandler and InstantiatePrefabHandler

2. Add null check for assetFolderPath in CreatePrefabHandler
   - Validates required parameter before processing

3. Remove duplicate SaveAssets call in SavePrefabHandler
   - Was called at line 52 and again at line 68, now only at end

4. Add asset path validation with AssetPathToGUID check
   - Better error messages distinguishing path errors from load failures
   - Applied to InstantiatePrefabHandler and OpenPrefabHandler
